requirements.txt
streamlit==1.39.0
langchain==0.3.2
langchain-community==0.3.1
langgraph==0.2.35
tavily-python==0.5.0
chromadb==0.5.5
pydantic==2.8.2
python-dotenv==1.0.1
unstructured==0.15.12
unstructured[html]==0.15.12
unstructured[pdf]==0.15.12
httpx==0.27.0
requests==2.32.3